fixed quadratic spline path through nodes, also decided to go easy on the C-style casts.

+++ VevaciousPlusPlus/include/BasicFunctions/SimplePolynomial.hpp
@@ -123,7 +123,7 @@ namespace VevaciousPlusPlus
          ++whichPower )
     {
       coefficientVector[ whichPower - 1 ]
-      = ( (double)whichPower * integralCoefficients[ whichPower ] );
+      = ( whichPower * integralCoefficients[ whichPower ] );
     }
   }
